+++ Sources/Sentry/SentryOptions.m
@@ -144,6 +144,13 @@
     } else {
         self.sessionTrackingIntervalMillis = [@30000 unsignedIntValue];
     }
+
+    if (nil != [options objectForKey:@"attachStacktrace"]) {
+        self.attachStacktrace =
+            [NSNumber numberWithBool:[[options objectForKey:@"attachStacktrace"] boolValue]];
+    } else {
+        self.attachStacktrace = @NO;
+    }
 }
 
 @end