+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- feat: Attach DebugMeta to Events #545
 - fix: Duplicate symbol for SentryMeta #549
 - feat: Set SUPPORTS_MACCATALYST to YES explicitly #547